+++ src/Linna/App/Views/HomeView.php
@@ -11,7 +11,6 @@ declare(strict_types=1);
 
 namespace Linna\App\Views;
 
-use Linna\App\Models\HomeModel;
 use Linna\App\Templates\HtmlTemplate;
 use Linna\Authentication\Authentication;
 use Linna\Mvc\View;
@@ -24,13 +23,12 @@ class HomeView extends View
     /**
      * Constructor.
      *
-     * @param HomeModel      $model
      * @param HtmlTemplate   $htmlTemplate
      * @param Authentication $login
      */
-    public function __construct(HomeModel $model, HtmlTemplate $htmlTemplate, Authentication $login)
+    public function __construct(HtmlTemplate $htmlTemplate, Authentication $login)
     {
-        parent::__construct($model, $htmlTemplate);
+        parent::__construct($htmlTemplate);
 
         //merge data passed from model with login information
         $this->data = \array_merge($this->data, ['login' => $login->islogged(), 'userName' => $login->getLoginData()['user_name']]);